pkg-deps.json: drop the Python version field and update the redhat deps
As Python 3 is the only supported version the Python version field can
be dropped from pkg-deps.json.

In CentOS 8 and in the newer Fedora releases the Python 3 packages are
named python3-<pkgname>, just like in Ubuntu, so we can drop most of the
renames

The httpretty, unittest2 and contextlib2 packages are renamed to the
empty string as they are not available on CentOS 8. They are test
dependencies, and since we are not running the unit tests when building
the RPM package they can be safely ignored.